analysis_crew/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.0
azure-functions==1.18.0
azure-identity==1.15.0
azure-search-documents==11.4.0b8
azure-storage-blob==12.18.3
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contextlib2==21.6.0
contourpy==1.2.0
crewai==0.28.8
crewai-tools==0.0.15
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docstring-parser==0.15
docx2txt==0.8
effdet==0.4.1
email_validator==2.1.1
embedchain==0.1.99
emoji==2.10.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.0
filelock==3.13.3
filetype==1.2.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.45.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.22.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
iniconfig==2.0.0
install==1.3.5
instructor==0.5.2
iopath==0.1.10
ipykernel==6.29.0
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.44
langchain-experimental==0.0.57
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.37
layoutparser==0.3.4
lxml==5.1.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.15.1
openai==1.14.3
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pgvector==0.2.4
pikepdf==8.14.0
pillow==10.2.0
pinecone-client==3.1.0
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.1
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
pydeck==0.8.1b0
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.18
Pympler==1.0.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==3.1.2
pypdf==3.17.4
pypdfium2==4.28.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pypyodbc==1.3.6
pyreadline3==3.4.1
pyright==1.1.356
pysbd==0.3.4
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.1.2
python-magic==0.4.27
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
ratelimiter==1.2.0.post0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
schema==0.7.5
scipy==1.12.0
selenium==4.19.0
semver==3.0.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.36.3
streamlit==1.32.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.39.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==4.3.1
ujson==5.9.0
unstructured==0.11.6
unstructured-client==0.17.0
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
urllib3==2.2.0
uvicorn==0.28.0
validators==0.24.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.18.1
